convert fullbright to flatshade + lower tex limit
somehow I was able to load a player model with a size of 720x720, and in the hl engine binary I can see 0x8000 is the limit before you get "texture too big". However, I consistently crash when using other models with that size of texture. Maybe because there are multiple of them. Lowering back to 512x512 to be safe.